Let's not move ModuleDefinitionSource
