+++ gradle-spaghetti-plugin/src/main/java/com/prezi/spaghetti/gradle/internal/AbstractSpaghettiTask.java
@@ -5,8 +5,8 @@ import com.prezi.spaghetti.bundle.ModuleBundle;
 import com.prezi.spaghetti.bundle.ModuleBundleSet;
 import com.prezi.spaghetti.definition.ModuleConfiguration;
 import com.prezi.spaghetti.definition.internal.ModuleConfigurationParser;
-import com.prezi.spaghetti.ast.ModuleDefinitionSource;
-import com.prezi.spaghetti.ast.internal.DefaultModuleDefinitionSource;
+import com.prezi.spaghetti.definition.ModuleDefinitionSource;
+import com.prezi.spaghetti.definition.internal.DefaultModuleDefinitionSource;
 import com.prezi.spaghetti.internal.DeprecationNagger;
 import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.internal.ConventionTask;
