+++ spaghetti-api/src/main/java/com/prezi/spaghetti/ast/Location.java
@@ -1,5 +1,7 @@
 package com.prezi.spaghetti.ast;
 
+import com.prezi.spaghetti.definition.ModuleDefinitionSource;
+
 /**
  * A location of a node in the original source.
  */
