+++ spaghetti-api/src/main/java/com/prezi/spaghetti/ast/ModuleNode.java
@@ -1,5 +1,7 @@
 package com.prezi.spaghetti.ast;
 
+import com.prezi.spaghetti.definition.ModuleDefinitionSource;
+
 public interface ModuleNode extends NamedNode, DocumentedNode, AnnotatedNode, MethodContainer {
 	String getAlias();
 	QualifiedNodeSet<ImportNode> getImports();
