+++ spaghetti-core/src/main/java/com/prezi/spaghetti/ast/internal/parser/ModuleParser.java
@@ -2,7 +2,7 @@ package com.prezi.spaghetti.ast.internal.parser;
 
 import com.google.common.collect.Lists;
 import com.prezi.spaghetti.ast.FQName;
-import com.prezi.spaghetti.ast.ModuleDefinitionSource;
+import com.prezi.spaghetti.definition.ModuleDefinitionSource;
 import com.prezi.spaghetti.ast.internal.DefaultFQName;
 import com.prezi.spaghetti.ast.internal.DefaultImportNode;
 import com.prezi.spaghetti.ast.internal.DefaultMethodNode;
