+++ spaghetti-core/src/main/java/com/prezi/spaghetti/definition/internal/ModuleConfigurationParser.java
@@ -5,9 +5,8 @@ import com.google.common.base.Throwables;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
-import com.prezi.spaghetti.ast.ModuleDefinitionSource;
+import com.prezi.spaghetti.definition.ModuleDefinitionSource;
 import com.prezi.spaghetti.ast.ModuleNode;
-import com.prezi.spaghetti.ast.internal.DefaultModuleDefinitionSource;
 import com.prezi.spaghetti.ast.internal.parser.AstParserException;
 import com.prezi.spaghetti.ast.internal.parser.MissingTypeResolver;
 import com.prezi.spaghetti.ast.internal.parser.ModuleParser;
@@ -16,7 +15,6 @@ import com.prezi.spaghetti.ast.internal.parser.TypeResolver;
 import com.prezi.spaghetti.bundle.ModuleBundle;
 import com.prezi.spaghetti.bundle.ModuleBundleSet;
 import com.prezi.spaghetti.definition.ModuleConfiguration;
-import com.prezi.spaghetti.definition.internal.DefaultModuleConfiguration;
 
 import java.io.IOException;
 import java.util.Collection;
