+++ spaghetti-core/src/test/groovy/com/prezi/spaghetti/config/ModuleConfigurationParserTest.groovy
@@ -1,6 +1,6 @@
 package com.prezi.spaghetti.config
 
-import com.prezi.spaghetti.ast.internal.DefaultModuleDefinitionSource
+import com.prezi.spaghetti.definition.internal.DefaultModuleDefinitionSource
 import com.prezi.spaghetti.ast.internal.parser.AstParserException
 import com.prezi.spaghetti.definition.internal.ModuleConfigurationParser
 import spock.lang.Specification
