+++ spaghetti-core/src/test/groovy/com/prezi/spaghetti/obfuscation/SymbolCollectVisitorTest.groovy
@@ -1,6 +1,6 @@
 package com.prezi.spaghetti.obfuscation
 
-import com.prezi.spaghetti.ast.internal.DefaultModuleDefinitionSource
+import com.prezi.spaghetti.definition.internal.DefaultModuleDefinitionSource
 import com.prezi.spaghetti.ast.internal.parser.ModuleParser
 import com.prezi.spaghetti.ast.internal.parser.TypeResolver
 import com.prezi.spaghetti.obfuscation.internal.SymbolCollectVisitor
